requirements.txt
attrs>=19.1.0
boto3>=1.9.176
cs>=2.7
requests-exoscale-auth>=1.0
requests>=2.22.0
toml>=0.10.0
